feat(ui): Nạp/Xả rename + backfill 3 ngày, sheet khuyến nghị dinh dưỡng, theme sáng/tối, fix parse dấu phẩy cân nặng
- Đổi tên nút: "🍱 Ghi món ăn" → "⚡ Nạp", "🏃 Vận động" → "🔥 Xả"
- Bỏ nút "🍽️ Ăn thêm (kcal)" và modal nhập calo tay
- Fix parse dấu phẩy thập phân từ bàn phím VI/DE (hàm parseDecimal)
- Thêm giao diện Sáng/Tối qua theme toggle trong Cài đặt
- Sheet "Khuyến nghị hàng ngày" khi bấm pin Năng lượng chính
- Đổi màu biểu đồ Lịch sử: Dinh dưỡng (xanh dương), Năng lượng (cam)
- Tiêu đề mục Cài đặt to/đậm hơn
- Ghi vận động cho ngày quá khứ (3 ngày gần nhất: hôm qua/kia/kìa)
- i18n đầy đủ VI/EN/DE cho toàn bộ tính năng mới

Kết quả: 503/503 test PASS, tsc sạch, eslint sạch

+++ src/domain/food/customFoodInput.ts
@@ -1,4 +1,5 @@
 import type { FoodItem, Nutrition, PortionUnit } from '../../types/food';
+import { parseDecimal } from '../../lib/units';
 
 // Pure form-state → domain-object logic for the "add a custom food" flow in
 // FoodLogModal. No I/O, no state — unit-tested. The modal only calls
@@ -97,8 +98,11 @@ export function resetNutritionForUnitChange(
 
 // Tolerant numeric parse: blank/whitespace/NaN → 0 (never throws, never
 // produces NaN downstream — the form must stay usable even half-filled).
+// Uses parseDecimal (not raw parseFloat) so a comma-decimal keyboard (VI/DE
+// locale, e.g. "12,5") isn't silently truncated to "12" — same fix as the
+// weight-input bug from Wave 1.
 function parseNum(s: string): number {
-  const n = parseFloat(s);
+  const n = parseDecimal(s);
   return isNaN(n) ? 0 : n;
 }
 
@@ -162,7 +166,7 @@ function scalePer100gToPerServing(n: Nutrition, servingWeightG: number): Nutriti
 // throw or silently mis-scale every macro/micro). Used to gate the save button.
 export function isValidCustomFoodInput(input: CustomFoodInput): boolean {
   if (input.name.trim().length === 0) return false;
-  const kcal = parseFloat(input.energyKcal);
+  const kcal = parseDecimal(input.energyKcal);
   if (isNaN(kcal) || kcal < 0) return false;
   if (input.portionUnit === 'pack' || input.portionUnit === 'capsule') {
     const servingWeightG = parseNum(input.servingWeightG);
